+++ tests/rheoQCM.py
@@ -63,6 +63,7 @@ from modules import QCM as QCM
 from modules.MatplotlibWidget import MatplotlibWidget
 
 import _version
+print('Version: {}'.format(_version.__version__))
 
 
 if UIModules.system_check() == 'win32': # windows
@@ -1989,6 +1990,7 @@ class QCMApp(QMainWindow):
                 print('File with wrong fromat!')
                 return
             else:
+                # remove some k
                 for key, val in settings.items():
                     self.settings[key] = val
 
@@ -2191,6 +2193,7 @@ class QCMApp(QMainWindow):
         logger.info(self.data_saver.path) 
         # re-initiate file
         self.data_saver.init_file(self.data_saver.path, settings=self.settings, t0=self.settings['dateTimeEdit_reftime'])
+
         # enable widgets
         self.enable_widgets(
             'pushButton_runstop_disable_list',
@@ -2200,8 +2203,9 @@ class QCMApp(QMainWindow):
         # clear spectra_fit items
         self.clr_spectra_fit()
 
-        # clear plainTextEdit
-        self.ui.plainTextEdit_settings_sampledescription.clear()
+        # in most condition, you do not want to clear the sample description
+        # # clear plainTextEdit
+        # self.ui.plainTextEdit_settings_sampledescription.clear()
 
 
     def on_triggered_del_menu(self, chn_name, plt_harms, sel_idx_dict, mode, marks):
@@ -2331,6 +2335,7 @@ class QCMApp(QMainWindow):
         set self.data_saver.path and lineEdit_datafilestr
         '''
         self.data_saver.path = fileName
+        self.tempPath = '' # init class file path information
         self.ui.lineEdit_datafilestr.setText(fileName)
 
 
@@ -2907,11 +2912,12 @@ class QCMApp(QMainWindow):
                 [self.active['harm']],
                 fs=[fit_result['v_fit']['cen_rec']['value']], # fs
                 gs=[fit_result['v_fit']['wid_rec']['value']], # gs = half_width
+                ps=[fit_result['v_fit']['amp_rec']['value']], # gs = half_width
             )
             # update mpl_plt12
             self.update_mpl_plt12()
 
-        #TODO add results to textBrowser_spectra_fit_result
+        # add results to textBrowser_spectra_fit_result
         self.ui.plainTextEdit_spectra_fit_result.setPlainText(self.peak_tracker.fit_result_report())
 
 
@@ -3395,7 +3401,7 @@ class QCMApp(QMainWindow):
     def get_harm_data_by_typestr(self, typestr, chn_name, harm, mark=False, unit_t=None, unit_temp=None):
         '''
         get data of a single harmonics from data_saver by given type (str) and harm (str)
-        str: 'df' ('delf_exps'), 'dfn', 'mdf', 'mdfn', 'dg' ('delg_exps'), 'dgn', 'f', 'g', 'temp', 't'
+        str: 'df' ('delf_exps'), 'dfn', 'mdf', 'mdfn', 'dg' ('delg_exps'), 'dgn', 'f', 'g', 'p', 'temp', 't'
         return: harm data
         '''
 
@@ -3435,6 +3441,10 @@ class QCMApp(QMainWindow):
             data = self.data_saver.get_marked_harm_col_from_list_column(chn_name, harm, 'fs', deltaval=False, norm=False, mark=mark)
         elif 'g' == typestr: # get g
             data = self.data_saver.get_marked_harm_col_from_list_column(chn_name, harm, 'gs', deltaval=False, norm=False, mark=mark)
+        elif 'p' == typestr: # get p
+            data = self.data_saver.get_marked_harm_col_from_list_column(chn_name, harm, 'ps', deltaval=False, norm=False, mark=mark)
+        elif 'gp' == typestr: # get gp
+            data = self.data_saver.get_marked_harm_col_from_list_column(chn_name, harm, 'ps', deltaval=False, norm=False, mark=mark) * self.data_saver.get_marked_harm_col_from_list_column(chn_name, harm, 'gs', deltaval=False, norm=False, mark=mark)
         elif 'D' == typestr: # get delta D
             # get delg first
             data = self.data_saver.get_marked_harm_col_from_list_column(chn_name, harm, 'gs', deltaval=False, norm=False, mark=mark)
@@ -3481,6 +3491,10 @@ class QCMApp(QMainWindow):
             data = self.data_saver.get_list_column_to_columns_marked_rows(chn_name, 'fs', mark=mark, dropnanmarkrow=False, deltaval=False, norm=False)
         elif 'g' == typestr: # get g
             data = self.data_saver.get_list_column_to_columns_marked_rows(chn_name, 'gs', mark=mark, dropnanmarkrow=False, deltaval=False, norm=False)
+        elif 'p' == typestr: # get p
+            data = self.data_saver.get_list_column_to_columns_marked_rows(chn_name, 'ps', mark=mark, dropnanmarkrow=False, deltaval=False, norm=False)
+        elif 'gp' == typestr: # get p
+            data = self.data_saver.get_list_column_to_columns_marked_rows(chn_name, 'ps', mark=mark, dropnanmarkrow=False, deltaval=False, norm=False) * self.data_saver.get_list_column_to_columns_marked_rows(chn_name, 'gs', mark=mark, dropnanmarkrow=False, deltaval=False, norm=False)
         elif 't' == typestr: # get t
             data = self.data_saver.get_t_marked_rows(chn_name, dropnanmarkrow=False, unit=unit_t)
         elif 'temp' == typestr: # get temp
@@ -4562,7 +4576,7 @@ class QCMApp(QMainWindow):
                         else: # upper layers
                             qcm_queue = qcm_df_layer.loc[[ind], :].copy() # as a dataframe
                             # get prop
-                            drho, grho_refh, phi, dlam_refh, err = self.qcm.solve_single_queue_to_prop(nh, qcm_queue, calctype, bulklimit=bulklimit)
+                            drho, grho_refh, phi, dlam_refh, err = self.qcm.solve_single_queue_to_prop(nh, qcm_queue, calctype=calctype, bulklimit=bulklimit)
 
                             prop_dict[ind][n].update(drho=drho, grho=grho_refh, phi=phi, n=refh)
                 else: 
@@ -6649,6 +6663,7 @@ class QCMApp(QMainWindow):
         f, G, B = {}, {}, {}
         fs = {} # peak centers
         gs = {} # dissipations hwhm
+        ps = {} # 
         curr_time = {}
         curr_temp = {}
         marks = [0 for _ in harm_list] # 'samp' and 'ref' chn test the same harmonics
@@ -6657,6 +6672,7 @@ class QCMApp(QMainWindow):
             f[chn_name], G[chn_name], B[chn_name] = {}, {}, {}
             fs[chn_name] = []
             gs[chn_name] = []
+            ps[chn_name] = []
             curr_temp[chn_name] = None
 
             self.reading = True
@@ -6757,6 +6773,7 @@ class QCMApp(QMainWindow):
                     # save data to fs and gs
                     fs[chn_name].append(fit_result['v_fit']['cen_rec']['value']) # fs
                     gs[chn_name].append(fit_result['v_fit']['wid_rec']['value'] ) # gs = half_width
+                    ps[chn_name].append(fit_result['v_fit']['amp_rec']['value'] ) # 
                     logger.info(cen_rec_freq) 
                     logger.info(cen_rec_G) 
 
@@ -6776,6 +6793,7 @@ class QCMApp(QMainWindow):
                     # save data to fs and gs
                     fs[chn_name].append(np.nan) # fs
                     gs[chn_name].append(np.nan) # gs = half_width
+                    ps[chn_name].append(np.nan) # gs = half_width
                     # clear lines
                     getattr(self.ui, 'mpl_sp' + harm).clr_lines(l_list=['lGfit', 'lBfit', 'lPfit', 'lsp', 'srec'])
 
@@ -6814,7 +6832,7 @@ class QCMApp(QMainWindow):
         if self.spectra_refresh_modulus() == 0: # check if to save by intervals
             self.writing = True
             # save raw
-            self.data_saver.dynamic_save(chn_name_list, harm_list, t=curr_time, temp=curr_temp, f=f, G=G, B=B, fs=fs, gs=gs, marks=marks)
+            self.data_saver.dynamic_save(chn_name_list, harm_list, t=curr_time, temp=curr_temp, f=f, G=G, B=B, fs=fs, gs=gs, ps=ps, marks=marks)
 
             # save data (THIS MIGHT MAKE THE PROCESS SLOW)
             self.data_saver.save_data()
@@ -6874,6 +6892,7 @@ class QCMApp(QMainWindow):
             # scan harmonics (1, 3, 5...)
             fs = []
             gs = []
+            ps = []
 
             self.reading = True
 
@@ -6896,6 +6915,7 @@ class QCMApp(QMainWindow):
                 # save data to fs and gs
                 fs.append(fit_result['v_fit']['cen_rec']['value']) # fs
                 gs.append(fit_result['v_fit']['wid_rec']['value'] ) # gs = half_width
+                ps.append(fit_result['v_fit']['amp_rec']['value'] ) #
 
                 # update lsp
                 factor_span = self.peak_tracker.get_output(key='factor_span', chn_name=chn_name, harm=harm)
@@ -6965,10 +6985,10 @@ class QCMApp(QMainWindow):
                 temp = self.data_saver.get_temp_C_from_raw(chn_name, queue_id)
                 marks = [0 for _ in harm_list] # 'samp' and 'ref' chn have the same harmonics
 
-                self.data_saver._save_queue_data([chn_name], harm_list, queue_id=queue_id, t={chn_name: t}, temp={chn_name: temp}, fs={chn_name: fs}, gs={chn_name: gs}, marks=marks)
+                self.data_saver._save_queue_data([chn_name], harm_list, queue_id=queue_id, t={chn_name: t}, temp={chn_name: temp}, fs={chn_name: fs}, gs={chn_name: gs}, ps={chn_name: ps}, marks=marks)
             else:
                 # save scan data to file fitting data in data_saver
-                self.data_saver.update_refit_data(chn_name, queue_id, harm_list, fs=fs, gs=gs)
+                self.data_saver.update_refit_data(chn_name, queue_id, harm_list, fs=fs, gs=gs, ps=ps)
 
             # plot data
             self.update_mpl_plt12()