+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
-## [0.19.0] - 2019-11-05
+
+## [0.19.2] - 2020-01-21
+
+### Added
+
+### Changed
+
+- NOTE: Save raw data (Gp & Bp) in S instead of mS
+
+### Fixed
+
+- Fix not reading user settings file bug.
+- Fix file name did not clear in main class after UI reset.
+
+### Removed
+
+## [0.19.1] - 2019-11-05
 
 ### Added